+++ scripts/scripts.js
@@ -45,22 +45,6 @@ export function loadScript(url, callback) {
   return script;
 }
 
-/*
-  * Returns the environment type based on the hostname.
-*/
-export function getEnvType(hostname = window.location.hostname) {
-  const fqdnToEnvType = {
-    'bevespi.com': 'live',
-    'www.bevespi.com': 'live',
-    'franklin.bevespi.com': 'live',
-    'main--bevespi--hlxsites.hlx.page': 'preview',
-    'main--bevespi--hlxsites.hlx.live': 'live',
-    'analytics--bevespi--hlxsites.hlx.page': 'preview',
-    'analytics--bevespi--hlxsites.hlx.live': 'live',
-  };
-  return fqdnToEnvType[hostname] || 'dev';
-}
-
 /**
  * Builds all synthetic blocks in a container element.
  * @param {Element} main The container element
